@@ -1,35 +1,35 @@
 {
-	"name": "pokemon-unfuse",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"dev": "vite dev",
-		"build": "vite build",
-		"preview": "vite preview",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
-	},
-	"devDependencies": {
-		"@poppanator/sveltekit-svg": "^3.0.0",
-		"@savvywombat/tailwindcss-grid-areas": "^3.0.1",
-		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/kit": "^1.15.0",
-		"@svelteuidev/composables": "^0.10.0",
-		"@types/node": "^18.15.11",
-		"autoprefixer": "^10.4.14",
-		"daisyui": "^2.51.5",
-		"postcss": "^8.4.21",
-		"svelte": "^3.58.0",
-		"svelte-check": "^3.1.4",
-		"tailwind-scrollbar": "^3.0.0",
-		"tailwindcss": "^3.3.1",
-		"tslib": "^2.5.0",
-		"typescript": "^5.0.3",
-		"vite": "^4.2.1",
-		"zod": "^3.21.4"
-	},
-	"type": "module",
-	"dependencies": {
-		"@sveltejs/adapter-vercel": "^2.4.1"
-	}
+  "name": "pokemon-unfuse",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "dev": "vite dev",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
+  },
+  "devDependencies": {
+    "@poppanator/sveltekit-svg": "^4.1.3",
+    "@savvywombat/tailwindcss-grid-areas": "^3.1.0",
+    "@sveltejs/adapter-auto": "^3.0.1",
+    "@sveltejs/kit": "^2.0.6",
+    "@svelteuidev/composables": "^0.15.4",
+    "@types/node": "^20.10.6",
+    "autoprefixer": "^10.4.16",
+    "daisyui": "^4.4.24",
+    "postcss": "^8.4.32",
+    "svelte": "5.0.0-next.27",
+    "svelte-check": "^3.6.2",
+    "tailwind-scrollbar": "^3.0.5",
+    "tailwindcss": "^3.4.0",
+    "tslib": "^2.6.2",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.10",
+    "zod": "^3.22.4"
+  },
+  "type": "module",
+  "dependencies": {
+    "@sveltejs/adapter-vercel": "^4.0.3"
+  }
 }
