@@ -4,22 +4,26 @@
 	import type { PokemonItem, PokemonList } from '../types'
 	import SelectImage from './SelectImage.svelte'
 
-	export let selected: PokemonItem | undefined = undefined
-	export let pokemonPromisses: Promise<PokemonList>
-	export let title: string
-	export let exclude: PokemonItem | undefined = undefined
+	let { selected, pokemonPromises, title, exclude } = $props<{
+		selected?: PokemonItem
+		pokemonPromises: Promise<PokemonList>
+		title: string
+		exclude?: PokemonItem
+	}>()
 
-	let pokemons: Awaited<typeof pokemonPromisses>['results'] = []
-	let dropdown: HTMLDivElement
-	let innerHeight: number
-	let scrollY: number
+	let pokemons = $state<Awaited<typeof pokemonPromises>['results']>([])
+	let dropdown = $state<HTMLDivElement>()
+	let innerHeight = $state<number>(0)
+	let scrollY = $state<number>()
 
-	$: pokemonPromisses.then((items) => {
-		pokemons = items.results
+	$effect(() => {
+		pokemonPromises.then((items) => {
+			pokemons = items.results
+		})
 	})
 
-	let isOpen = false
-	let filter = ''
+	let isOpen = $state(false)
+	let filter = $state('')
 
 	const close = () => {
 		isOpen = false
@@ -29,11 +33,17 @@
 		isOpen = !isOpen
 	}
 
-	$: dropdownBottom = scrollY || true ? innerHeight - dropdown?.getBoundingClientRect().bottom > 320 : true
+	const dropdownBottom = $derived(
+		scrollY || true ? innerHeight - (dropdown?.getBoundingClientRect().bottom ?? 0) > 320 : true
+	)
 
-	$: filteredPokemons = pokemons
-		.filter((item) => item !== exclude)
-		.filter((item) => item.name.toLowerCase().includes(filter.toLowerCase()))
+	const filteredPokemons = $derived(
+		pokemons.filter((item) => item !== exclude).filter((item) => item.name.toLowerCase().includes(filter.toLowerCase()))
+	)
+
+	const autofocus = (node: HTMLInputElement) => {
+		node.focus()
+	}
 </script>
 
 <svelte:window bind:innerHeight bind:scrollY />
@@ -45,7 +55,7 @@
 	class="dropdown {dropdownBottom ? 'dropdown-bottom' : 'dropdown-top'} w-60 text-center"
 >
 	<label for="btn-{title}" class="label-text">{title}</label>
-	{#await pokemonPromisses}
+	{#await pokemonPromises}
 		<button disabled id="btn-{title}" class="btn btn-primary w-full">Loading Pokemons...</button>
 	{:then}
 		<button id="btn-{title}" on:click={toggle} class="btn btn-primary w-full">
@@ -57,8 +67,7 @@
 			transition:slide
 			class="absolute dropdown-content z-10 w-full grid grid-flow-row menu p-2 shadow bg-slate-900 text-slate-100 rounded-box gap-2"
 		>
-			<!-- svelte-ignore a11y-autofocus -->
-			<input class="input input-bordered input-primary" autofocus bind:value={filter} placeholder="Filter" />
+			<input class="input input-bordered input-primary" use:autofocus bind:value={filter} placeholder="Filter" />
 			<ul class="overflow-y-auto h-80 scrollbar-none">
 				{#each filteredPokemons as item}
 					<li class="flex w-full {item.url === selected?.url ? 'btn-primary' : ''}">
