@@ -1,5 +1,5 @@
 import adapter from '@sveltejs/adapter-vercel';
-import { vitePreprocess } from '@sveltejs/kit/vite';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -10,6 +10,10 @@ const config = {
 			runtime: 'edge'
 		}),
 		inlineStyleThreshold: 60000,
+	},
+
+	compilerOptions: {
+		runes: true,
 	}
 };
 
